+++ ProductManagement.Application.Contract/ProductViewModel.cs
@@ -5,7 +5,7 @@
         public long Id { get; set; }
         public string? Name { get; set; }
         public string? Picture { get; set; }
-        public decimal PriceInIran { get; set; }
+        public long PriceInIran { get; set; }
         public string CreationDate { get; set; }
     }
 }
